VID-2142 Some fixes after code review meeting
* Moved modules into new namespace
* Cleaned up some unused code
